Encode Atlas
made a function in segmentation that takes an image and has a hardcoded dict that maps colors to region names (which will be uploaded based on the atlas by roberto). It generates a dictionary of regions mapped to coordinates.
There is a tester function which uses the encoding and the image splitting function to split apart a dummy image.